14: Renamed CalendarProfileScreen to CalendarsSelectionScreen
